+++ samplesheets/io.py
@@ -8,6 +8,7 @@ import time
 
 from .models import Investigation, Study, Assay, GenericMaterial, Protocol, \
     Process
+from .rendering import SampleSheetTableBuilder
 
 
 # Local constants
@@ -392,7 +393,7 @@ def get_inv_paths(zip_file):
 # iRODS Utils ------------------------------------------------------------------
 
 
-def get_irods_dirs(container):
+def get_base_dirs(container):
     """
     Return iRODS directory structure for an Investigation or a part of it
     :param container: an Investigation, Study or Assay object
@@ -423,3 +424,39 @@ def get_irods_dirs(container):
         dirs.append(assay.get_dir(include_study=True))
 
     return dirs
+
+
+def get_assay_dirs(assay):
+    """
+    Return iRODS directory structure under an assay
+    :param assay: Assay object
+    :return: List
+    """
+    # TODO: This is just an example, implement this for real
+
+    # NOTE: This is still the best and fastest way to figure out which samples
+    #       belong in a study
+
+    tb = SampleSheetTableBuilder()
+    tables = tb.build_study(assay.study)
+
+    # TODO: Add common helpers for getting sample column / position
+    sample_pos = 0
+
+    for c in tables['study']['top_header']:
+        if c['value'] == 'Sample':
+            break
+
+        else:
+            sample_pos += c['colspan']
+
+    assay_table = tables['assays'][assay.get_name()]
+    ret = []
+
+    for row in assay_table['table_data']:
+        val = row[sample_pos]['value']
+
+        if val not in ret:
+            ret.append(val)
+
+    return ret