+++ samplesheets/models.py
@@ -239,6 +239,11 @@ class Study(BaseSampleSheet):
         """Return simple printable name for study"""
         return self.title if self.title else self.identifier
 
+    def get_display_name(self):
+        """Return display name for study"""
+        # return self.title
+        return self.title.strip('.').title() if self.title else self.identifier
+
     def get_dir(self):
         """Return directory name for study"""
         return 'study_' + str(self.omics_uuid)
@@ -392,8 +397,12 @@ class Assay(BaseSampleSheet):
     # Custom row-level functions
 
     def get_name(self):
-        """Return simple idenfitying name for Assay"""
+        """Return simple idenfitying name for assay"""
         return ''.join(str(self.file_name)[2:].split('.')[:-1])
+
+    def get_display_name(self):
+        """Return display name for assay"""
+        return ' '.join(s for s in self.get_name().split('_')).title()
 
     def get_dir(self, include_study=False):
         """
@@ -401,7 +410,6 @@ class Assay(BaseSampleSheet):
         :param include_study: Include parent study directory in string (bool)
         :return: String
         """
-
         return '{}assay_{}'.format(
             (self.study.get_dir() + '/') if include_study else '',
             self.omics_uuid)